Migrate getWeekdaySequence
